- regression test should run - cleanup workspace scripts - task workspaces:setup should be run before running workspace scripts - cloud examples should not infinite loop
