+++ .gitignore
@@ -5,8 +5,8 @@ node_modules
 .task/
 *.swp
 .DS_Store
-optic-snapshots/
-output/
+/optic-snapshots
+/output
 /target
 /workspaces/diff-engine-wasm/engine/browser
 /workspaces/diff-engine-wasm/engine/build
