+++ taskfiles/regression-tests/Taskfile.yml
@@ -9,7 +9,7 @@ tasks:
     deps: [clean]
     vars:
       API_PROJECT_DIR: '{{default "optic-snapshots" .API_PROJECT_DIR}}'
-      NUM_INTERACTIONS: '{{default 1 .NUM_INTERACTIONS}}'
+      NUM_INTERACTIONS: '{{default 10000 .NUM_INTERACTIONS}}'
       OUTPUT_DIR: '{{default "output" .OUTPUT_DIR}}'
       OPTIC_RUST_DIFF_ENGINE: '{{default "true" .OPTIC_RUST_DIFF_ENGINE}}'
     env:
@@ -27,22 +27,21 @@ tasks:
       - rm -rfv {{.API_PROJECT_DIR}}/.optic/captures/ccc/diffs/*
       - cd {{.API_PROJECT_DIR}} && node ../workspaces/local-cli/bin/run daemon:stop
       - cd {{.API_PROJECT_DIR}} && node ../workspaces/local-cli/bin/run spec
+      - cargo build
       - >
         node
-        workspaces/snapshot-tests/build/e2e/index.js
+        workspaces/snapshot-tests/build/regressions/index.js
         {{.OUTPUT_DIR}}/{{.NUM_INTERACTIONS}}
         {{.API_PROJECT_DIR}}
         {{.NUM_INTERACTIONS}}
-      - cat {{.OUTPUT_DIR}}/{{.NUM_INTERACTIONS}}/*
+      - ls -lah {{.OUTPUT_DIR}}/{{.NUM_INTERACTIONS}}/*
       - ls -lah {{.API_PROJECT_DIR}}/.optic/captures/ccc/diffs/*
-      - cat {{.API_PROJECT_DIR}}/.optic/captures/ccc/diffs/**/diff-engine-output.log
 
   on-failure:
     vars:
       API_PROJECT_DIR: '{{default "optic-snapshots" .API_PROJECT_DIR}}'
     cmds:
       - ls -lah {{.API_PROJECT_DIR}}/.optic/captures/ccc/diffs/*
-      - cat {{.API_PROJECT_DIR}}/.optic/captures/ccc/diffs/**/diff-engine-output.log
 
   example-inputs:
     vars:
