--- workspace-scripts/notify.sh
@@ -1,4 +1,0 @@
-MESSAGE=$1
-osascript -e "display notification \"${MESSAGE}\""
-# say "$MESSAGE"
-
