+++ workspaces/spectacle-shared/.gitignore
@@ -1,0 +1,2 @@
+build
+node_modules
