+++ workspaces/spectacle-shared/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "@useoptic/spectacle-shared",
+  "version": "10.0.1",
+  "scripts": {
+    "ws:build": "tsc -b --verbose",
+    "ws:clean": "rm -rf build"
+  },
+  "repository": {
+    "type": "git",
+    "url": "ssh://git@github.com/useoptic/optic-package.git",
+    "directory": "workspaces/spectacle-shared"
+  },
+  "main": "build/index.js",
+  "types": "build/index.d.ts",
+  "files": [
+    "/build"
+  ],
+  "dependencies": {
+    "@useoptic/spectacle": "10.0.1"
+  },
+  "devDependencies": {}
+}
