+++ workspaces/spectacle-shared/src/index.ts
@@ -1,0 +1,1 @@
+export * from './local-cli/client';
