+++ workspaces/spectacle-shared/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "extends": "../config/tsconfig.json",
+  "compilerOptions": {
+    "outDir": "build",
+    "resolveJsonModule": true,
+    "rootDir": "src"
+  },
+  "include": ["src/**/*"],
+  "exclude": ["test"]
+}
