+++ workspaces/ui-v2/src/spectacle-implementations/public-examples.tsx
@@ -1,26 +1,23 @@
 import React, { useCallback, useEffect, useState } from 'react';
 import { Switch, useParams, useRouteMatch } from 'react-router-dom';
-import { Provider as BaseUrlProvider } from '../optic-components/hooks/useBaseUrl';
-import {
-  IBaseSpectacle,
-  IForkableSpectacle,
-  IOpticContext,
-  makeSpectacle,
-  SpectacleInput,
-} from '@useoptic/spectacle';
-import { DocumentationPages } from '../optic-components/pages/docs/DocumentationPage';
+import { Provider as BaseUrlProvider } from '<src>/optic-components/hooks/useBaseUrl';
+import { DocumentationPages } from '<src>/optic-components/pages/docs/DocumentationPage';
 import { SpectacleStore } from './spectacle-provider';
-import { Loading } from '../optic-components/loaders/Loading';
-import { DiffReviewEnvironments } from '../optic-components/pages/diffs/ReviewDiffPages';
-import { InMemoryOpticContextBuilder } from '@useoptic/spectacle/build/in-memory';
-import { CapturesServiceStore } from '../optic-components/hooks/useCapturesHook';
-import { ChangelogPages } from '../optic-components/pages/changelog/ChangelogPages';
+import { Loading } from '<src>/optic-components/loaders/Loading';
+import { DiffReviewEnvironments } from '<src>/optic-components/pages/diffs/ReviewDiffPages';
+import {
+  InMemoryOpticContextBuilder,
+  InMemorySpectacle,
+  InMemoryBaseSpectacle,
+} from '@useoptic/spectacle/build/in-memory';
+import { CapturesServiceStore } from '<src>/optic-components/hooks/useCapturesHook';
+import { ChangelogPages } from '<src>/optic-components/pages/changelog/ChangelogPages';
 import {
   AppConfigurationStore,
   OpticAppConfig,
-} from '../optic-components/hooks/config/AppConfiguration';
-import { useOpticEngine } from '../optic-components/hooks/useOpticEngine';
-import { AsyncStatus } from '../types';
+} from '<src>/optic-components/hooks/config/AppConfiguration';
+import { useOpticEngine } from '<src>/optic-components/hooks/useOpticEngine';
+import { AsyncStatus } from '<src>/types';
 import { ConfigRepositoryStore } from '<src>/optic-components/hooks/useConfigHook';
 
 const appConfig: OpticAppConfig = {
@@ -71,9 +68,7 @@ export default function PublicExamples() {
     return <div>something went wrong</div>;
   }
 
-  //@ts-ignore
-  data.isOG = true;
-
+  //@SYNC public-examples.tsx cloud-viewer.tsx local-cli.tsx
   return (
     <AppConfigurationStore config={appConfig}>
       <SpectacleStore spectacle={data}>
@@ -104,42 +99,6 @@ export interface InMemorySpectacleDependencies {
 
 export type InMemorySpectacleDependenciesLoader = () => Promise<InMemorySpectacleDependencies>;
 
-export class InMemorySpectacle
-  implements IForkableSpectacle, InMemoryBaseSpectacle {
-  private spectaclePromise: ReturnType<typeof makeSpectacle>;
-
-  constructor(
-    public readonly opticContext: IOpticContext,
-    public samples: any[]
-  ) {
-    this.spectaclePromise = makeSpectacle(opticContext);
-  }
-
-  async fork(): Promise<IBaseSpectacle> {
-    const opticContext = await InMemoryOpticContextBuilder.fromEventsAndInteractions(
-      this.opticContext.opticEngine,
-      [...(await this.opticContext.specRepository.listEvents())],
-      this.samples,
-      'example-session'
-    );
-    return new InMemorySpectacle(opticContext, [...this.samples]);
-  }
-
-  async mutate<Result, Input = {}>(options: SpectacleInput<Input>) {
-    const spectacle = await this.spectaclePromise;
-    return spectacle.queryWrapper<Result, Input>(options);
-  }
-
-  async query<Result, Input = {}>(options: SpectacleInput<Input>) {
-    const spectacle = await this.spectaclePromise;
-    return spectacle.queryWrapper<Result, Input>(options);
-  }
-}
-
-export interface InMemoryBaseSpectacle extends IBaseSpectacle {
-  samples: any[];
-  opticContext: IOpticContext;
-}
 //@SYNC: useInMemorySpectacle
 export function useInMemorySpectacle(
   loadDependencies: InMemorySpectacleDependenciesLoader
